# Conflicts:
#	src/main/java/com/mercadolibro/dto/InvoiceDTO.java
#	src/main/java/com/mercadolibro/dto/InvoiceSearchDTO.java
#	src/main/java/com/mercadolibro/entity/Invoice.java
#	src/main/java/com/mercadolibro/service/impl/InvoiceRequestServiceImpl.java

+++ src/main/java/com/mercadolibro/dto/CategorySalesDTO.java
@@ -1,8 +1,6 @@
 package com.mercadolibro.dto;
 
 public interface CategorySalesDTO {
-
     String getCategory_name();
-    Integer getSales();
-
+    Long getSales();
 }
